web-scraper-python/requirements.txt
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
feedparser==6.0.11
Flask==3.0.3
python-dotenv==1.0.1
idna==3.7
requests==2.32.3
soupsieve==2.5
urllib3==2.2.2
